+++ fueoni_ver2/lib/models/arguments.dart
@@ -1,0 +1,17 @@
+class CreationRoomArguments {
+  int? roomId;
+
+  CreationRoomArguments({required this.roomId});
+}
+
+class LoadingRoomArguments {
+  int? roomId;
+
+  LoadingRoomArguments({required this.roomId});
+}
+
+class SearchRoomArguments {
+  int? roomId;
+
+  SearchRoomArguments({required this.roomId});
+}
