--- fueoni_ver2/lib/services/database/loading_room_services.dart
@@ -1,5 +1,0 @@
-class LoadingRoomArguments {
-  int? roomId;
-
-  LoadingRoomArguments({required this.roomId});
-}
