+++ fueoni_ver2/lib/services/room_search/game_monitor_service.dart
@@ -1,0 +1,14 @@
+import 'package:firebase_database/firebase_database.dart';
+
+class GameMonitorService {
+  void monitorGameStart(int? roomId, Function(bool) onGameStartChanged) {
+    DatabaseReference gameStartRef =
+        FirebaseDatabase.instance.ref('games/$roomId/settings/gameStart');
+
+    gameStartRef.onValue.listen((event) {
+      if (event.snapshot.exists && event.snapshot.value == true) {
+        onGameStartChanged(true);
+      }
+    });
+  }
+}
